Simplify interference graph setup
